simplify dispatcher around the idea of a single channel
That way, all objects go away and it will be nothing more than a task
around a channel.
